restructured app + added CatSourceComponent3FGL and its route + dropdown in CatHelpComponent
