+++ src/app/app.routes.ts
@@ -1,8 +1,9 @@
 import { provideRouter, RouterConfig } from '@angular/router';
-import { MapViewComponent } from './map-view/map-view.component';
-import { CatViewComponent } from './cat-view/cat-view.component';
-import { CatHelpComponent } from './cat-view/cat-help/cat-help.component';
-import { CatSource2FHLComponent } from './cat-view/cat-source/cat-source-2fhl/cat-source-2fhl.component';
+import { MapViewComponent } from './views/map-view/map-view.component';
+import { CatViewComponent } from './views/cat-view/cat-view.component';
+import { CatHelpComponent } from './widgets/cat-help/cat-help.component';
+import { CatSource2FHLComponent } from './widgets/cat-source/cat-source-2fhl/cat-source-2fhl.component';
+import { CatSource3FGLComponent } from './widgets/cat-source/cat-source-3fgl/cat-source-3fgl.component';
 
 const routes: RouterConfig = [
 
@@ -17,6 +18,10 @@ const routes: RouterConfig = [
       {
         path: '2fhl',
         component: CatSource2FHLComponent
+      },
+      {
+        path: '3fgl',
+        component: CatSource3FGLComponent
       },
       {
         path: '',
