--- src/app/cat-view/cat-search/cat-search.component.html
@@ -1,4 +1,0 @@
-<p>
-  This is CatSearchComponent
-  <button (click)="onClick()">Click to view 2FHL catalog</button>
-</p>
