+++ src/app/views/cat-view/cat-view.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ROUTER_DIRECTIVES } from '@angular/router';
 import { SwitchViewComponent } from '../switch-view/switch-view.component';
-import { CatSearchComponent } from './cat-search/cat-search.component';
+import { CatSearchComponent } from '../../widgets/cat-search/cat-search.component';
 
 @Component({
   moduleId: module.id,
