+++ src/app/views/map-view/map-view.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { SwitchViewComponent } from '../switch-view/switch-view.component';
 import {Router} from '@angular/router';
-import {MapComponent} from './map/map.component';
-import {AboutButtonComponent} from './about-button/about-button.component';
+import {MapComponent} from '../../widgets/map/map.component';
+import {AboutButtonComponent} from '../../widgets/about-button/about-button.component';
 
 @Component({
   moduleId: module.id,
