+++ src/app/views/switch-view/index.ts
