+++ src/app/views/switch-view/switch-view.component.html
