+++ src/app/views/switch-view/switch-view.component.spec.ts
