+++ src/app/widgets/cat-search/cat-search.component.html
@@ -1,0 +1,10 @@
+<p>
+  This is CatSearchComponent
+
+  <select (change)="onSelectChange($event.target.value)">
+    <option value="">Choose a catalog</option>
+    <option value="3fgl">3FGL</option>
+    <option value="2fhl">2FHL</option>
+  </select>
+  
+</p>
