+++ src/app/widgets/cat-source/cat-source-3fgl/cat-source-3fgl.component.spec.ts
@@ -1,0 +1,13 @@
+/* tslint:disable:no-unused-variable */
+
+import { By }           from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
+import { addProviders, async, inject } from '@angular/core/testing';
+import { CatSource3FGLComponent } from './cat-source-3fgl.component';
+
+describe('Component: CatSource3FGL', () => {
+  it('should create an instance', () => {
+    let component = new CatSource3FGLComponent();
+    expect(component).toBeTruthy();
+  });
+});
