+++ src/app/widgets/cat-source/cat-source-3fgl/cat-source-3fgl.component.ts
@@ -1,0 +1,17 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  selector: 'cat-source-3fgl',
+  templateUrl: 'cat-source-3fgl.component.html',
+  styleUrls: ['cat-source-3fgl.component.css']
+})
+export class CatSource3FGLComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+    console.log("Routing to CatSource3FGLComponent...");
+  }
+
+}
