+++ src/app/widgets/cat-source/cat-source-3fgl/index.ts
@@ -1,0 +1,1 @@
+export * from './cat-source-3fgl.component';
