+++ src/app/widgets/map/map.component.spec.ts
