+++ src/app/widgets/map/map.component.ts
